+++ dkg/pedersen/dkg_struct.go
